Camera: use an empty app id in access dialog in case of empty app info
Sending an app id that doesn't resolve to valid app info (e.g. a desktop
file is not found) will automatically reject camera request, because the
backend is not able to verify whether the app (based on app id) runs in
the background. This happens for example when you start Firefox in GNOME
using Alt + F2 shortcut, where we get firefox" as app id from cgroups
and since it doesn't find opened application window for "firefox", it
automatically rejects the request. Using an empty id fixes this problem.